Update block1.svg
I kept the viewBox in there, but needed to make sure the length is longer than the svg image appears so that the hover effect for the drawbridge will work well while hovering underneath the drawbridge. I also added the text: "click here" for mobile users and any device that doesn't support hover.